+++ core/qc/__init__.py
@@ -3,3 +3,5 @@ from .localization import localize_qc_report
 from .models import CurveQCStatistics, QCFinding, QCReport, QC_SCHEMA
 
 __all__ = ["LasQCEngine", "localize_qc_report", "CurveQCStatistics", "QCFinding", "QCReport", "QC_SCHEMA"]
+
+from .artifacts import QCPersistenceResult, QCReportArtifactService